requirements.txt
anyio==3.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
mpmath==1.3.0
python-telegram-bot==20.3
requests==2.30.0
sniffio==1.3.0
sympy==1.11.1
urllib3==2.0.2
